Safari-609.1.11.1/LayoutTests/http/tests/eventloop/queue-task-across-frames-expected.txt

This tests the order by which tasks are scheduled across documents of similar origins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS logs.join(", ") is "1, 2, 3, 4, 5, 6"
PASS successfullyParsed is true

TEST COMPLETE
